# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Gen-C Scripts API
  version: 1.0.0
  description: |
    OpenAPI specification for key backend surfaces used by the dashboard.
    The documented endpoints power script CRUD flows in the dashboard and the
    YouTube idea seed generator. Responses follow the existing `{ success: ... }`
    envelope used by the legacy fetch helpers.
servers:
  - url: https://api.gencapp.pro
    description: Production
  - url: http://localhost:5001
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Collection:
      type: object
      additionalProperties: true
      required: [id, title, userId]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        userId: { type: string }
        videoCount: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CollectionsResponse:
      type: object
      required: [success]
      additionalProperties: true
      properties:
        success:
          type: boolean
          enum: [true]
        collections:
          type: array
          items: { $ref: '#/components/schemas/Collection' }
    MoveVideoRequest:
      type: object
      additionalProperties: false
      required: [videoId, toCollectionId]
      properties:
        videoId: { type: string }
        fromCollectionId: { type: string }
        toCollectionId: { type: string }
    CopyVideoRequest:
      type: object
      additionalProperties: false
      required: [videoId, toCollectionId]
      properties:
        videoId: { type: string }
        fromCollectionId: { type: string }
        toCollectionId: { type: string }
    UpdateCollectionRequest:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
    VideosResponse:
      type: object
      additionalProperties: true
      required: [success, videos]
      properties:
        success:
          type: boolean
          enum: [true]
        videos:
          type: array
          items:
            type: object
            additionalProperties: true
        totalCount: { type: number }
        lastDocId: { type: string }
...
    ErrorResponse:
      type: object
      required: [success, error]
      additionalProperties: true
      properties:
        success:
          type: boolean
          enum: [false]
        error: { type: string }
        debug:
          type: object
          description: Optional debug payload returned for internal errors.
...
    BasicSuccessResponse:
      type: object
      required: [success]
      additionalProperties: true
      properties:
        success:
          type: boolean
          enum: [true]
...
paths:
...
  /api/collections:
    get:
      summary: List collections for the authenticated user.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Collections fetched successfully.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CollectionsResponse' }
        '401':
          description: Authentication required.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/collections/move-video:
    post:
      summary: Move a video between collections.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MoveVideoRequest' }
      responses:
        '200':
          description: Video moved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BasicSuccessResponse' }
        '400':
          description: Invalid payload.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Authentication required.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/collections/copy-video:
    post:
      summary: Copy a video to another collection.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CopyVideoRequest' }
      responses:
        '200':
          description: Video copied.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BasicSuccessResponse' }
        '400':
          description: Invalid payload.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Authentication required.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/collections/update:
    patch:
      summary: Update a collection's metadata.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCollectionRequest' }
      responses:
        '200':
          description: Collection updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BasicSuccessResponse' }
        '400':
          description: Invalid payload.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Authentication required.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/collections/delete:
    delete:
      summary: Delete a collection by id.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Collection deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BasicSuccessResponse' }
        '401':
          description: Authentication required.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/videos/collection:
    post:
      summary: List videos for a collection.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              properties:
                collectionId: { type: string }
                videoLimit: { type: number }
                lastDocId: { type: string }
      responses:
        '200':
          description: Videos returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VideosResponse' }
        '401':
          description: Authentication required.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...